マイグレーション修正

+++ src/database/migrations/2026_01_12_081304_create_contacts_table.php
@@ -15,7 +15,7 @@ class CreateContactsTable extends Migration
     {
         Schema::create('contacts', function (Blueprint $table) {
             $table->id();
-            $table->integer('category_id');
+            $table->bigInteger('category_id');
             $table->string('first_name');
             $table->string('last_name');
             $table->tinyInteger('gender')->comment('1:男性 2:女性 3:その他');
